pyproject.toml
[project]
name = "LLMToys"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.12,<3.14.1"
dependencies = [
    "langchain>=1.2.15",
    "langchain-core>=1.2.15",
    "langchain-ollama>=1.1.0",
    "llmlingua>=0.2.2",
    "pandas>=2.3.3",
    "pydantic>=2.11.0",
    "pyyaml>=6.0.3",
    "pymssql>=2.3.13",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.49",
    "sqlglot>=30.6.0",
    "torch==2.11.0",
    "torch-geometric>=2.7.0",
    "torchaudio==2.11.0",
    "torchvision==0.26.0",
    "transformers>=5.0.0",
    "vllm>=0.20.0,<0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black (>=26.3.1,<27.0.0)",
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
# sentencepiece 0.2.1, pulled transitively by vllm, emits these SWIG warnings on Python 3.13.
filterwarnings = [
    "ignore:builtin type SwigPyPacked has no __module__ attribute:DeprecationWarning",
    "ignore:builtin type SwigPyObject has no __module__ attribute:DeprecationWarning",
    "ignore:builtin type swigvarlink has no __module__ attribute:DeprecationWarning",
]

[tool.uv]
package = false
override-dependencies = ["transformers>=5.0.0"]

[tool.uv.sources]
torch = { index = "pytorch-cu130" }
torchvision = { index = "pytorch-cu130" }
torchaudio = { index = "pytorch-cu130" }

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true
